refactor constants, add environment vars for webhooks & chatwoot marker
