+++ samanthas_telegram_bot/auxil/log_and_notify.py
@@ -4,9 +4,8 @@ from dotenv import load_dotenv
 from telegram import Bot, Update
 from telegram.constants import ParseMode
 
-from samanthas_telegram_bot.auxil.constants import ADMIN_CHAT_ID
+from samanthas_telegram_bot.auxil.constants import ADMIN_CHAT_ID, CALLER_LOGGING_STACK_LEVEL
 from samanthas_telegram_bot.auxil.escape_for_markdown import escape_for_markdown
-from samanthas_telegram_bot.data_structures.constants import CALLER_LOGGING_STACK_LEVEL
 from samanthas_telegram_bot.data_structures.enums import LoggingLevel
 
 load_dotenv()
