+++ samanthas_telegram_bot/data_structures/models.py
@@ -7,7 +7,7 @@ from typing import Optional, TypedDict
 
 from samanthas_telegram_bot.api_clients.auxil.constants import DataDict
 from samanthas_telegram_bot.api_clients.auxil.enums import SmalltalkTestStatus
-from samanthas_telegram_bot.data_structures.constants import WEEKDAYS
+from samanthas_telegram_bot.auxil.constants import WEEKDAYS
 from samanthas_telegram_bot.data_structures.enums import AgeRangeType
