+++ sample.env
@@ -5,8 +5,12 @@ BOT_OWNER_USERNAME=owner_telegram_username
 CHATWOOT_ACCOUNT_ID=0
 CHATWOOT_API_TOKEN=abcde1235fghij
 CHATWOOT_INBOX_ID=0
+CHATWOOT_MARKER_FOR_MESSAGE_TO_BE_SENT_TO_OPERATOR=send_this
 LOGGING_LEVEL=INFO
 SMALLTALK_TOKEN=a1b2c3d4e5f6
 SMALLTALK_TEST_ID=z1y2x3w4v5
 SPEAKING_CLUB_COORDINATOR_USERNAME=some_telegram_username
 TELEGRAM_WEBHOOK_SECRET_TOKEN=telegram_webhook_1234567SomeToken
+WEBHOOK_URL_PREFIX=https://domain.tld
+WEBHOOK_PATH_FOR_CHATWOOT=relative-path-1
+WEBHOOK_PATH_FOR_TELEGRAM=relative-path-2
